requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
langchain==0.1.6
langchain-google-genai==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
openpyxl==3.1.2